Frst step: extract the http api client into a separate package
