+++ cli/bzk/variant.go
@@ -17,7 +17,7 @@ func listVariantsCommand(cmd *cli.Cmd) {
 	})
 
 	cmd.Action = func() {
-		client, err := NewClient(checkServerURI(*bzkUri))
+		client, err := NewClient()
 		if err != nil {
 			log.Fatal(err)
 		}
@@ -43,7 +43,7 @@ func variantLogCommand(cmd *cli.Cmd) {
 		Desc: "the variant id",
 	})
 	cmd.Action = func() {
-		client, err := NewClient(checkServerURI(*bzkUri))
+		client, err := NewClient()
 		if err != nil {
 			log.Fatal(err)
 		}
